Title: 사각형 채우기 3, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240322/사각형 채우기 3/rectangle-fill-3.cpp
@@ -2,7 +2,7 @@
 #define MAXN 1000
 using namespace std;
 
-int dp[MAXN + 1] = {1, 2, 7, 0, };
+long long dp[MAXN + 1] = {1, 2, 7, 0, };
 
 int main() {
     int n;
